Pack Particle's id and lifetime into one four-byte slot
id and lifetime are both ushort now and sit next to each other at the head
of the second block, which takes sizeof(Particle) from 84 to 80 with no
padding anywhere - verified against the compiler, every offset in the
comments included. Neither is near its ceiling: 65535 ticks is twenty-two
minutes, and a spark id is reused only after 32768 conversions, by which
time the sparks of the first are two seconds dead.

nextSparkId() keeps its guarantee in sixteen bits - the top bit set, so a
spark id can still never be the 0 that every other particle carries.

The narrowing at the call sites is written out rather than left implicit,
or MSVC reports C4244 at /W3 in twenty-nine places that were quiet before.
Measured with gcc -Wconversion over the fifteen files that spawn
particles: one warning before, one after, and that one predates this.

+++ Blocks5/src/cannon.cpp
@@ -78,7 +78,7 @@ void Cannon::onUpdate()
 		ParticleSystem::Particle p;
 
 		// smoke
-		p.lifetime = random(25, 50);
+		p.lifetime = static_cast<ushort>(random(25, 50));
 		p.damping = 0.99f;
 		p.gravity = 0.005f;
 		p.positionOnTexture = Vec2b(0, 0);
@@ -175,7 +175,7 @@ bool Cannon::fire()
 	// fire/smoke forward and to the sides
 	for(int i = 0; i < 100; i++)
 	{
-		p.lifetime = random(5, 10);
+		p.lifetime = static_cast<ushort>(random(5, 10));
 		p.damping = 0.99f;
 		p.gravity = 0.005f;
 		p.positionOnTexture = Vec2b(64, 0);
